Integer zoom percent and input-owned software cursor.
Zoom is discrete 50–400% (step 10, 100 = exact 1:1) with integer
pan and transforms, so identity blit stays reachable after zooming.
Software cursor paint is owned by the input thread; compositor paths
only lift the glyph on damage and mark dirty, avoiding stale 32x32
underlay squares when the mouse moves over a redrawing window.

+++ input_headless.c
@@ -50,17 +50,17 @@ int resize_buffer(struct Client *c, int dx, int dy)
 
 struct Client *pick_client(int x, int y)
 {
-	float wx, wy;
+	int wx, wy;
 	struct Client *c;
 	struct Client *picked = NULL;
 
-	screen_to_world(&server, (float)x, (float)y, &wx, &wy);
+	screen_to_world(&server, x, y, &wx, &wy);
 	c = server.clients;
 	while (c) {
 		uint32_t ww = c->world_w ? c->world_w : c->width;
 		uint32_t wh = c->world_h ? c->world_h : c->height;
-		if (wx >= (float)c->x && wx < (float)(c->x + ww) &&
-		    wy >= (float)c->y && wy < (float)(c->y + wh)) {
+		if (wx >= (int)c->x && wx < (int)(c->x + ww) &&
+		    wy >= (int)c->y && wy < (int)(c->y + wh)) {
 			picked = c;
 			break;
 		}
